fix: mobileSection component화
